Fixed the issue where the translated KB is deleted...
